Created all endpoints, added validation, made corresponding changes.

+++ bank-system/src/main/java/com/mybanksystem/transaction/model/entity/FlatAmountProvisionTransaction.java
@@ -6,22 +6,23 @@ import com.mybanksystem.transaction.model.enumeration.TransactionType;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 import javax.persistence.Transient;
+import java.math.BigDecimal;
 
 @Entity
 @DiscriminatorValue("flatamount")
 public class FlatAmountProvisionTransaction extends Transaction {
     @Transient
-    double provision;
+    BigDecimal provision;
     public FlatAmountProvisionTransaction() {
     }
 
-    public FlatAmountProvisionTransaction(Account accountFrom, Account accountTo, Double amount, String description, TransactionType type, Double provision) {
+    public FlatAmountProvisionTransaction(Account accountFrom, Account accountTo, BigDecimal amount, String description, TransactionType type, BigDecimal provision) {
         super(accountFrom, accountTo, amount, description, type);
         this.provision = provision;
     }
 
     @Override
-    public double getProvision() {
+    public BigDecimal getProvision() {
 
         return provision;
         //return getBank().getFlatFeeAmount();
